Fyyur: Large application structure
