+++ projects/01_fyyur/starter_code/app/__init__.py
@@ -1,0 +1,27 @@
+from flask import Flask
+from flask_sqlalchemy import SQLAlchemy
+from flask_bootstrap import Bootstrap
+from flask_moment import Moment
+from config import config
+# from flask.logging import create_logger
+
+db = SQLAlchemy()
+moment = Moment()
+bootstrap = Bootstrap()
+# log = create_logger()
+
+
+def create_app(config_name):
+    app = Flask(__name__)
+    app.config.from_object(config[config_name])
+    config[config_name].init_app(app)
+
+    db.init_app(app)
+    moment.init_app(app)
+    bootstrap.init_app(app)
+    # log.init_app(app)
+
+    from .main import main as main_blueprint
+    app.register_blueprint(main_blueprint)
+
+    return app
