+++ projects/01_fyyur/starter_code/app/static/css/main.quickfix.css
