+++ projects/01_fyyur/starter_code/app/static/fonts/fontawesome-webfont.ttf
