+++ projects/01_fyyur/starter_code/app/static/img/front-splash.jpg
