+++ projects/01_fyyur/starter_code/app/static/js/plugins.js
