+++ projects/01_fyyur/starter_code/app/static/js/script.js
