+++ projects/01_fyyur/starter_code/app/templates/forms/edit_venue.html
@@ -1,10 +1,10 @@
-{% extends 'layouts/main.html' %} {% block title %}New Venue{% endblock %} {%
+{% extends 'layouts/main.html' %} {% block title %}Edit Venue{% endblock %} {%
 block content %}
 <div class="form-wrapper">
-  <form method="post" class="form">
+  <form class="form" method="post" action="/venues/{{venue.id}}/edit">
     <h3 class="form-heading">
-      List a new venue
-      <a href="{{ url_for('index') }}" title="Back to homepage"
+      Edit venue <em>{{ venue.name }}</em>
+      <a href="{{ url_for('main.index') }}" title="Back to homepage"
         ><i class="fa fa-home pull-right"></i
       ></a>
     </h3>
@@ -45,7 +45,11 @@ block content %}
       {{ form.facebook_link(class_ = 'form-control', placeholder='http://',
       autofocus = true) }}
     </div>
-    <input type="submit" value="Create Venue" class="btn btn-primary btn-lg" />
+    <input
+      type="submit"
+      value="Edit Venue"
+      class="btn btn-primary btn-lg btn-block"
+    />
   </form>
 </div>
 {% endblock %}
