+++ projects/01_fyyur/starter_code/app/templates/forms/new_venue.html
@@ -1,10 +1,10 @@
-{% extends 'layouts/main.html' %} {% block title %}Edit Venue{% endblock %} {%
+{% extends 'layouts/main.html' %} {% block title %}New Venue{% endblock %} {%
 block content %}
 <div class="form-wrapper">
-  <form class="form" method="post" action="/venues/{{venue.id}}/edit">
+  <form method="post" class="form">
     <h3 class="form-heading">
-      Edit venue <em>{{ venue.name }}</em>
-      <a href="{{ url_for('index') }}" title="Back to homepage"
+      List a new venue
+      <a href="{{ url_for('main.index') }}" title="Back to homepage"
         ><i class="fa fa-home pull-right"></i
       ></a>
     </h3>
@@ -45,11 +45,7 @@ block content %}
       {{ form.facebook_link(class_ = 'form-control', placeholder='http://',
       autofocus = true) }}
     </div>
-    <input
-      type="submit"
-      value="Edit Venue"
-      class="btn btn-primary btn-lg btn-block"
-    />
+    <input type="submit" value="Create Venue" class="btn btn-primary btn-lg" />
   </form>
 </div>
 {% endblock %}
