+++ projects/01_fyyur/starter_code/app/templates/layouts/main.html
@@ -53,9 +53,9 @@
         <div class="collapse navbar-collapse">
           <ul class="nav navbar-nav">
             <li>
-              {% if (request.endpoint == 'venues') or
-                (request.endpoint == 'search_venues') or
-                (request.endpoint == 'show_venue') %}
+              {% if (request.endpoint == 'main.venues') or
+                (request.endpoint == 'main.search_venues') or
+                (request.endpoint == 'main.show_venue') %}
               <form class="search" method="post" action="/venues/search">
                 <input class="form-control"
                   type="search"
@@ -64,9 +64,9 @@
                   aria-label="Search">
               </form>
               {% endif %}
-              {% if (request.endpoint == 'artists') or
-                (request.endpoint == 'search_artists') or
-                (request.endpoint == 'show_artist') %}
+              {% if (request.endpoint == 'main.artists') or
+                (request.endpoint == 'main.search_artists') or
+                (request.endpoint == 'main.show_artist') %}
               <form class="search" method="post" action="/artists/search">
                 <input class="form-control"
                   type="search"
@@ -78,9 +78,9 @@
             </li>
           </ul>
           <ul class="nav navbar-nav">
-            <li {% if request.endpoint == 'venues' %} class="active" {% endif %}><a href="{{ url_for('venues') }}">Venues</a></li>
-            <li {% if request.endpoint == 'artists' %} class="active" {% endif %}><a href="{{ url_for('artists') }}">Artists</a></li>
-            <li {% if request.endpoint == 'shows' %} class="active" {% endif %}><a href="{{ url_for('shows') }}">Shows</a></li>
+            <li {% if request.endpoint == 'main.venues' %} class="active" {% endif %}><a href="{{ url_for('main.venues') }}">Venues</a></li>
+            <li {% if request.endpoint == 'main.artists' %} class="active" {% endif %}><a href="{{ url_for('main.artists') }}">Artists</a></li>
+            <li {% if request.endpoint == 'mani.shows' %} class="active" {% endif %}><a href="{{ url_for('main.shows') }}">Shows</a></li>
           </ul>
         </div><!--/.nav-collapse -->
       </div>
