+++ projects/01_fyyur/starter_code/app/templates/pages/show.html
