--- projects/01_fyyur/starter_code/error.log
@@ -1,1 +1,0 @@
-
