+++ projects/01_fyyur/starter_code/fyyur.py
@@ -1,0 +1,29 @@
+import os
+import sys
+import json
+import babel
+import dateutil.parser
+from app import create_app, db
+from flask_migrate import Migrate
+from app.models import Artist, Show, Venue
+
+
+app = create_app(os.getenv('FLASK_CONFIG') or 'default')
+migrate = Migrate(app, db)
+
+
+@app.shell_context_processor
+def make_shell_context():
+    return dict(db=db, Artist=Artist, Venue=Venue, Show=Show)
+
+
+def format_datetime(value, format='medium'):
+    date = dateutil.parser.parse(value)
+    if format == 'full':
+        format = "EEEE MMMM, d, y 'at' h:mma"
+    elif format == 'medium':
+        format = "EE MM, dd, y h:mma"
+    return babel.dates.format_datetime(date, format)
+
+
+app.jinja_env.filters['datetime'] = format_datetime
